# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Sample CRUD api for estuary db",
        "title": "Sample CRUD api for estuary db",
        "contact": {
            "name": "Me",
            "url": "http://me.com/terms.html",
            "email": "alvin@protocol.ai"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/contentdealproposallogs": {
            "get": {
                "description": "GetAllContentDealProposalLogs is a handler to get a slice of record(s) from content_deal_proposal_logs table in the estuary database",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ContentDealProposalLogs"
                ],
                "summary": "Get list of ContentDealProposalLogs",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "page requested (defaults to 0)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "number of records in a page  (defaults to 20)",
                        "name": "pagesize",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "db sort order column",
                        "name": "order",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.PagedResults"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/model.ContentDealProposalLogs"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.HTTPError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/api.HTTPError"
                        }
                    }
                }
            },
            "post": {
                "description": "add to add a single record to content_deal_proposal_logs table in the estuary database",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ContentDealProposalLogs"
                ],
                "summary": "Add an record to content_deal_proposal_logs table",
                "parameters": [
                    {
                        "description": "Add ContentDealProposalLogs",
                        "name": "ContentDealProposalLogs",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.ContentDealProposalLogs"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.ContentDealProposalLogs"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.HTTPError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/api.HTTPError"
                        }
                    }
                }
            }
        },
        "/contentdealproposallogs/{contentDealProposalLogsID}": {
            "get": {
                "description": "GetContentDealProposalLogs is a function to get a single record from the content_deal_proposal_logs table in the estuary database",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ContentDealProposalLogs"
                ],
                "summary": "Get record from table ContentDealProposalLogs by  contentDealProposalLogsID",
                "operationId": "contentDealProposalLogsID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id",
                        "name": "contentDealProposalLogsID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.ContentDealProposalLogs"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.HTTPError"
                        }
                    },
                    "404": {
                        "description": "ErrNotFound, db record for id not found - returns NotFound HTTP 404 not found error",
                        "schema": {
                            "$ref": "#/definitions/api.HTTPError"
                        }
                    }
                }
            },
            "put": {
                "description": "Update a single record from content_deal_proposal_logs table in the estuary database",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ContentDealProposalLogs"
                ],
                "summary": "Update an record in table content_deal_proposal_logs",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id",
                        "name": "contentDealProposalLogsID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update ContentDealProposalLogs record",
                        "name": "ContentDealProposalLogs",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.ContentDealProposalLogs"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.ContentDealProposalLogs"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.HTTPError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/api.HTTPError"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a single record from content_deal_proposal_logs table in the estuary database",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ContentDealProposalLogs"
                ],
                "summary": "Delete a record from content_deal_proposal_logs",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id",
                        "name": "contentDealProposalLogsID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content",
                        "schema": {
                            "$ref": "#/definitions/model.ContentDealProposalLogs"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.HTTPError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "api.HTTPError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 400
                },
                "message": {
                    "type": "string",
                    "example": "status bad request"
                }
            }
        },
        "api.PagedResults": {
            "type": "object",
            "properties": {
                "data": {},
                "page": {
                    "type": "integer"
                },
                "pageSize": {
                    "type": "integer"
                },
                "totalRecords": {
                    "type": "integer"
                }
            }
        },
...
        "model.ContentDealProposalLogs": {
            "type": "object",
            "properties": {
                "content": {
                    "description": "[ 1] content                                        INT8                 null: true   primary: false  isArray: false  auto: false  col: INT8            len: -1      default: []",
                    "$ref": "#/definitions/null.Int"
                },
                "createdAt": {
                    "description": "[ 9] created_at                                     TIMESTAMPTZ          null: true   primary: false  isArray: false  auto: false  col: TIMESTAMPTZ     len: -1      default: []",
                    "type": "string"
                },
                "deltaNodeUUID": {
                    "description": "[11] delta_node_uuid                                TEXT                 null: true   primary: false  isArray: false  auto: false  col: TEXT            len: -1      default: []",
                    "$ref": "#/definitions/null.String"
                },
                "id": {
                    "description": "[ 0] id                                             INT8                 null: false  primary: true   isArray: false  auto: false  col: INT8            len: -1      default: []",
                    "type": "integer"
                },
                "meta": {
                    "description": "[ 4] meta                                           TEXT                 null: true   primary: false  isArray: false  auto: false  col: TEXT            len: -1      default: []",
                    "$ref": "#/definitions/null.String"
                },
                "nodeInfo": {
                    "description": "[ 5] node_info                                      TEXT                 null: true   primary: false  isArray: false  auto: false  col: TEXT            len: -1      default: []",
                    "$ref": "#/definitions/null.String"
                },
                "requesterInfo": {
                    "description": "[ 6] requester_info                                 TEXT                 null: true   primary: false  isArray: false  auto: false  col: TEXT            len: -1      default: []",
                    "$ref": "#/definitions/null.String"
                },
                "requestingAPIKey": {
                    "description": "[ 7] requesting_api_key                             TEXT                 null: true   primary: false  isArray: false  auto: false  col: TEXT            len: -1      default: []",
                    "$ref": "#/definitions/null.String"
                },
                "signed": {
                    "description": "[ 3] signed                                         TEXT                 null: true   primary: false  isArray: false  auto: false  col: TEXT            len: -1      default: []",
                    "$ref": "#/definitions/null.String"
                },
                "systemContentDealProposalID": {
                    "description": "[ 8] system_content_deal_proposal_id                INT8                 null: true   primary: false  isArray: false  auto: false  col: INT8            len: -1      default: []",
                    "$ref": "#/definitions/null.Int"
                },
                "unsigned": {
                    "description": "[ 2] unsigned                                       TEXT                 null: true   primary: false  isArray: false  auto: false  col: TEXT            len: -1      default: []",
                    "$ref": "#/definitions/null.String"
                },
                "updatedAt": {
                    "description": "[10] updated_at                                     TIMESTAMPTZ          null: true   primary: false  isArray: false  auto: false  col: TIMESTAMPTZ     len: -1      default: []",
                    "type": "string"
                }
            }
        },
...
        "null.Int": {
            "type": "object",
            "properties": {
                "int64": {
                    "type": "integer"
                },
                "valid": {
                    "description": "Valid is true if Int64 is not NULL",
                    "type": "boolean"
                }
            }
        },
        "null.String": {
            "type": "object",
            "properties": {
                "string": {
                    "type": "string"
                },
                "valid": {
                    "description": "Valid is true if String is not NULL",
                    "type": "boolean"
                }
            }
        }
    }
}